services/api/requirements.txt
Django==4.0.4
gunicorn==20.1.0
psycopg2-binary==2.9.2
whitenoise==6.1.0
django-import-export==2.8.0
django-jazzmin==2.5.0
Pillow==9.1.0
djangorestframework==3.13.1